refactor: メッセージ内容の変更とエラーメッセージ格納場所の変更
