+++ native/app/lib/resources/messages_ja.tsx
@@ -1,0 +1,8 @@
+export const ERROR_MESSAGE = {
+  BAD_REQUEST: '入力内容に誤りがあります。',
+  UNAUTHORIZED: '権限がありません。アプリを再起動してください。',
+  PROCESS_FAILED: '正常に処理できませんでした。しばらく経ってからもう一度お試しください。',
+  SERVER_ERROR: 'サーバーエラーが発生しました。しばらく経ってからもう一度お試しください。',
+  TIMEOUT: '通信がタイムアウトしました。しばらく経ってからもう一度お試しください。',
+  UNEXPEXTED_ERROR: '予期せぬエラーが発生しました。しばらく経ってからもう一度お試しください。',
+};
