+++ native/app/screens/SignUp.tsx
@@ -10,7 +10,7 @@ import HeaderWithBackButton from '~/components/organisms/HeaderWithBackButton';
 import { emailValidation } from '~/lib/validation';
 import { SingUpForm } from '~/types/forms';
 import { Alert } from 'react-native';
-import { ERROR_MESSAGE } from '~~/constants/theme';
+import { generateErrorMessage } from '~/lib/util/ErrorUtil';
 
 const styles = StyleSheet.create({
   container: {
@@ -67,22 +67,6 @@ const SignUp = function SignUp(props: Props): ReactElement {
       ],
     );
 
-  const generateErrorMessage = function generateErrorMessage(code:number): string {
-    switch(code) {
-    case 400:
-      return ERROR_MESSAGE.BAD_REQUEST;
-    case 401:
-      return ERROR_MESSAGE.UNAUTHORIZED;
-    case 403 || 404 || 409:
-      return ERROR_MESSAGE.PROCESS_FAILED;
-    case 500 || 501 || 503:
-      return ERROR_MESSAGE.SERVER_ERROR;
-    case 504:
-      return ERROR_MESSAGE.TIMEOUT;
-    default:
-      return ERROR_MESSAGE.UNEXPEXTED_ERROR;
-    }
-  };
 
   const handleSubmit = React.useCallback(async () => {
     await signUpWithEmail(
