+++ native/constants/theme.ts
@@ -85,12 +85,3 @@ export const SOCIAL_BUTTON: ViewStyle ={
   width: 330,
   height: 42,
 };
-
-export const ERROR_MESSAGE = {
-  BAD_REQUEST: '入力内容に誤りがあります。',
-  UNAUTHORIZED: '権限がありません。アプリを再起動してください。',
-  PROCESS_FAILED: '正常に処理できませんでした。しばらく経ってからもう一度お試しください。',
-  SERVER_ERROR: 'サーバーエラーが発生しました。しばらく経ってからもう一度お試しください。',
-  TIMEOUT: '通信がタイムアウトしました。しばらく経ってからもう一度お試しください。',
-  UNEXPEXTED_ERROR: '予期せぬエラーが発生しました。',
-};
